feat: Fully migrate the repo from northwestern to grammes-go
While the repo had been forked, all the docs and more importantly, the module, referred to the northwestern version of the package,
which made it unusable in a module based build system.

This PR does the following:

   - Renames everything to use grammes-go rather than northwestern
   - Changes the mod to use go 1.16 - because of the backward compatibility of go, there is no reason not to
     specify 1.16
   - Updates some of the dependencies to their latest versions as otherwise the project will get too far
     behind. Only the websocket upgrade could really make any difference and that is only 1.4.0 -> 1.4.2,
     which in semver terms, should be safe.

This makes it look like quite a big PR, but the above changes are the only changes.